+++ src/components/AppHeader/AppHeader.tsx
@@ -1,8 +1,5 @@
 import React from 'react';
 import cn from 'classnames';
-import * as todosApi from '../../api/todos';
-import { Todo } from '../../types/Todo';
-import { ErrorMessages } from '../../types/ErrorMessages';
 
 interface Props {
   onSubmit: (title: string) => void;
@@ -11,11 +8,8 @@ interface Props {
   isInputActive: boolean;
   inputRef: React.RefObject<HTMLInputElement>;
   isAllTodosCompleted: boolean;
-  setOperatedTodo: React.Dispatch<React.SetStateAction<number[]>>;
-  todoData: Todo[];
-  setTodoData: React.Dispatch<React.SetStateAction<Todo[]>>;
-  setErrorMessage: (value: ErrorMessages) => void;
   shouldShowElement: boolean;
+  handleToggleAll: () => void;
 }
 
 export const AppHeader: React.FC<Props> = ({
@@ -25,64 +19,9 @@ export const AppHeader: React.FC<Props> = ({
   isInputActive,
   inputRef,
   isAllTodosCompleted,
-  setOperatedTodo,
-  todoData,
-  setTodoData,
-  setErrorMessage,
   shouldShowElement,
+  handleToggleAll,
 }) => {
-  const handleToggleAll = () => {
-    const newCompletedStatus = !isAllTodosCompleted;
-
-    const todosToUpdate = todoData.filter(
-      todo => todo.completed !== newCompletedStatus,
-    );
-
-    const todosInOperation = todosToUpdate.map(todo => todo.id);
-
-    setOperatedTodo(cur => [...cur, ...todosInOperation]);
-
-    Promise.allSettled(
-      todosToUpdate.map(todo => {
-        const patchedTodo = {
-          ...todo,
-          completed: !todo.completed,
-        };
-
-        return todosApi.patchTodo(todo.id, patchedTodo);
-      }),
-    )
-      .then(results => {
-        const patchedTodos = results.map((result, index) => {
-          if (result.status === 'fulfilled') {
-            return {
-              ...todosToUpdate[index],
-              completed: !todosToUpdate[index].completed,
-            };
-          }
-
-          return null;
-        });
-
-        if (patchedTodos.includes(null)) {
-          setErrorMessage(ErrorMessages.OnPatch);
-        }
-
-        setTodoData(cur => {
-          return cur.map(todo => {
-            const patched = patchedTodos.find(p => p && p.id === todo.id);
-
-            return patched ? patched : todo;
-          });
-        });
-      })
-      .finally(() =>
-        setOperatedTodo(cur =>
-          cur.filter(id => !todosInOperation.includes(id)),
-        ),
-      );
-  };
-
   return (
     <header className="todoapp__header">
       {shouldShowElement && (
@@ -96,7 +35,7 @@ export const AppHeader: React.FC<Props> = ({
         />
       )}
 
-      <form
+      <form // винестив  кромп
         onSubmit={event => {
           event.preventDefault();
           onSubmit(todoTitle.trim());
