+++ src/components/TodoItem/TodoItem.tsx
@@ -3,16 +3,15 @@
 import React, { useState } from 'react';
 import cn from 'classnames';
 import { Todo } from '../../types/Todo';
-import * as todosApi from '../../api/todos';
-import { ErrorMessages } from '../../types/ErrorMessages';
 
 interface Props {
   todo: Todo;
   isOverlayActive?: boolean;
   handleDelete?: (value: number) => void;
   handleSwitchStatus?: (value: number) => void;
-  setErrorMessage?: (value: ErrorMessages) => void;
-  handleUpdate?: (value: Todo) => void;
+  handleUpdate?: (text: string, value: Todo) => void;
+  isTodoEditing?: boolean;
+  setEditingTodoId?: (value: number | null) => void;
 }
 
 export const TodoItem: React.FC<Props> = ({
@@ -20,53 +19,17 @@ export const TodoItem: React.FC<Props> = ({
   isOverlayActive = true,
   handleDelete = () => {},
   handleSwitchStatus = () => {},
-  setErrorMessage = () => {},
   handleUpdate = () => {},
+  isTodoEditing = false,
+  setEditingTodoId = () => {},
 }) => {
   const { id, completed, title } = todo;
 
-  const [isEditing, setIsEditing] = useState(false);
   const [editingTitle, setEditingTitle] = useState(todo.title);
-  const [isSaving, setIsSaving] = useState(false);
 
   const handleSubmit = (event?: React.FormEvent<HTMLFormElement>) => {
-    if (event) {
-      event.preventDefault();
-    }
-
-    const normalizedTitle = editingTitle.trim();
-
-    if (!normalizedTitle) {
-      handleDelete(id);
-
-      return;
-    }
-
-    if (normalizedTitle === title) {
-      setIsEditing(false);
-
-      return;
-    }
-
-    setIsSaving(true);
-
-    const editedTodo = {
-      ...todo,
-      title: normalizedTitle,
-    };
-
-    todosApi
-      .patchTodo(id, editedTodo)
-      .then((response: unknown) => {
-        const patchedTodo = response as Todo;
-
-        handleUpdate(patchedTodo);
-        setIsEditing(false);
-      })
-      .catch(() => setErrorMessage(ErrorMessages.OnPatch))
-      .finally(() => {
-        setIsSaving(false);
-      });
+    event?.preventDefault();
+    handleUpdate(editingTitle, todo);
   };
 
   const handleBlur = () => {
@@ -75,7 +38,7 @@ export const TodoItem: React.FC<Props> = ({
 
   const handleKeyUp = (event: React.KeyboardEvent<HTMLInputElement>) => {
     if (event.key === 'Escape') {
-      setIsEditing(false);
+      setEditingTodoId(null);
       setEditingTitle(title);
     }
   };
@@ -92,12 +55,12 @@ export const TodoItem: React.FC<Props> = ({
         />
       </label>
 
-      {!isEditing ? (
+      {!isTodoEditing ? (
         <>
           <span
             data-cy="TodoTitle"
             className="todo__title"
-            onDoubleClick={() => setIsEditing(true)}
+            onDoubleClick={() => setEditingTodoId(id)}
           >
             {title}
           </span>
@@ -130,7 +93,7 @@ export const TodoItem: React.FC<Props> = ({
       <div
         data-cy="TodoLoader"
         className={cn('modal', 'overlay', {
-          'is-active': isOverlayActive || isSaving,
+          'is-active': isOverlayActive,
         })}
       >
         <div className="modal-background has-background-white-ter" />
