--- src/components/TodoList/index.ts
@@ -1,1 +1,0 @@
-export { TodoList } from './TodoList';
